Completed Classes
Tentative classes with parameters of member functions filled in.
Currently the system I have in mind is that the checkKeyword function
would see if the keyword applies to the item, and would call
performAction function, which would return an int, which is the
identifier for a certain action, which at this point we think we would
have to store all of which in an "event handler" type entity.