Debugs, cleanup, and misconception
Item is a parent of others, but a list of items cannot hold a child of
item (say, InventoryItems) pointers may help, but i can't rewrite
everything to use lists of item pointers in time.  sorry guys, we didn't
get it in on time. I'm submitting as is, then we'll have to try turning
a more complete draft in a day or two late as well, or just make sure to
rock the final version.